UHF-8508: Reduced duplicated code by adding abstract class for the selection widgets.
