+++ modules/hdbt_admin_tools/src/Plugin/Field/FieldWidget/SelectColorPaletteWidget.php
@@ -3,9 +3,9 @@
 namespace Drupal\hdbt_admin_tools\Plugin\Field\FieldWidget;
 
 use Drupal\Core\Field\FieldItemListInterface;
-use Drupal\Core\Field\Plugin\Field\FieldWidget\OptionsSelectWidget;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\hdbt_admin_tools\Form\SiteSettings;
+use Drupal\hdbt_admin_tools\SelectWidgetBase;
 
 /**
  * Plugin implementation of the 'color_palette_field_widget' widget.
@@ -20,7 +20,7 @@ use Drupal\hdbt_admin_tools\Form\SiteSettings;
  *   multiple_values = FALSE
  * )
  */
-class SelectColorPaletteWidget extends OptionsSelectWidget {
+class SelectColorPaletteWidget extends SelectWidgetBase {
 
   /**
    * {@inheritdoc}
@@ -34,68 +34,13 @@ class SelectColorPaletteWidget extends OptionsSelectWidget {
   ): array {
     $element = parent::formElement($items, $delta, $element, $form, $form_state);
 
-    $element['#type'] = 'select';
-    $element['#cardinality'] = $this->fieldDefinition->getFieldStorageDefinition()->getCardinality();
-    $element['#options'] = $this->getOptions($items->getEntity());
     $element['#default_value'] = !empty($this->getSelectedOptions($items))
       ? $this->getSelectedOptions($items)
       : SiteSettings::getColorPaletteDefaultValue();
-    $element['#theme'] = 'selection_widget';
     $element['#attached']['library'][] = 'hdbt_admin_tools/select_color_palette';
     $element['#attributes']['class'][] = 'select-color-palette';
 
     return $element;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function validateElement(
-    array $element,
-    FormStateInterface $form_state
-  ): void {
-
-    // Massage submitted form values.
-    // Drupal\Core\Field\WidgetBase::submit() expects values as
-    // an array of values keyed by delta first, then by column, while our
-    // widgets return the opposite.
-    if (is_array($element['#value'])) {
-      $values = array_values($element['#value']);
-    }
-    else {
-      $values = [$element['#value']];
-    }
-
-    // Filter out the '' option. Use a strict comparison, because
-    // 0 == 'any string'.
-    $index = array_search('', $values, TRUE);
-    if ($index !== FALSE) {
-      unset($values[$index]);
-    }
-
-    // Design field widget cannot handle multiple values.
-    $item[$element['#key_column']] = $element['#value'];
-    $form_state->setValueForElement($element, $item);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  protected function getEmptyLabel(): void {}
-
-  /**
-   * Get field name and convert it to a more suitable name for our needs.
-   *
-   * @return string
-   *   Returns field name.
-   */
-  protected function getFieldName(): string {
-    $field_name = $this->fieldDefinition->getName();
-    if ($field_name) {
-      $name = str_replace('field_', '', $field_name);
-      return str_replace('_', '-', $name);
-    }
-    return $this->fieldDefinition->getName();
   }
 
 }
